Add ChiralIsomorphism to score OpenBabel mappings
OpenBabel finds equivalent atoms based on a graph representation,
ignoring the 3D structure. For example, in chlorodifluoromethane,
which is officially not chiral, OpenBabel consideres the fluorine atoms
as equivalent (i.e. symmetric). This is not desirable, because from the
prespective of one fluorine atom, the carbon has H -> F -> Cl in a clockwise
direction, and from the prespective of the other fluorine atom, H -> F -> Cl
are in a counter-clockwise direction. The ChiralIsomorphism class scores
OpenBabel mappings according to angle differences between pairs of dihedral
angles involving equivalent atoms. It does not work for chlorodifluoromethane
because there are not enough atoms to calculate dihedral angles.